refactor(frontend): move check-locale script to scripts/locale/, changed usage to `npm run locale diff <base> <targets>`; added scripts/ to eslint pre-commit check scope; added locale jsons to eslint pre-commit check scope.
