+++ scripts/locale/index.js
@@ -3,17 +3,17 @@ const path = require("path");
 const subCommand = process.argv[2];
 
 if (!subCommand) {
-  console.error("Please specify a subcommand for version (check or bump)");
+  console.error("Please specify a subcommand for locale (diff)");
   process.exit(1);
 }
 
 try {
-  const args = process.argv.slice(4); // Get any additional arguments
+  const args = process.argv.slice(3); // Get any additional arguments
   require(path.join(__dirname, `${subCommand}.js`));
 } catch (error) {
   if (error.code === "MODULE_NOT_FOUND") {
     console.error(`Unknown subcommand: ${subCommand}`);
-    console.error("Available subcommands: check, bump");
+    console.error("Available subcommands: diff");
   } else {
     console.error(error);
   }
